requirements.txt
line-bot-sdk
feedparser
flask
mysql.connector